\name{MortHump-package}
\alias{MortHump-package}
\alias{MortHump}
\docType{package}
\title{
\packageTitle{MortHump}
}
\description{
\packageDescription{MortHump}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MortHump}
\packageIndices{MortHump}

}

\author{

\packageAuthor{MortHump}

Maintainer: \packageMaintainer{MortHump}

}

\references{

Remund A., Camarda C. and Riffe T. (2018) Analyzing the Young Adult Mortality Hump in R with MortHump. \emph{MPIDR Technical Report} TR-2018-3.

Remund, A., Riffe, T., & Camarda, C. (2018). A cause-of-death decomposition of young adult excess mortality. \emph{Demography}.

Camarda, C. G., Eilers, P. H. C., & Gampe, J. (2016). Sums of smooth exponentials to decompose complex series of counts. \emph{Statistical Modelling}.

Remund, A. (2015). Jeunesses vulnerables? Mesures, composantes et causes de la surmortalite des jeunes adultes. (PhD), Universite de Geneve, Geneve. \href{http://archive-ouverte.unige.ch/unige:73525}{[Free access]}

Camarda, C. G. (2008). Smoothing methods for the analysis of mortality development. (Doctorado en Ingenieria Matematica), Universidad Carlos III de Madrid, Madrid.

}

\keyword{Mortality hump -  P-spline - Causes of Death}
\seealso{
%\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
\pkg{MortalitySmooth}
\pkg{HMDHFDplus}
}
%\examples{
%~~ simple examples of the most important functions ~~
%}
